feedback from nick
